[Feat] [FE-47] Navigation 전체 수정
