+++ src/components/molecules/IconWrap/HomeIconWrap.tsx
@@ -7,7 +7,7 @@ import {
   Text,
   Dimensions,
 } from 'react-native';
-import { RootStackNavigationProp } from '@/types/NavigationTypes';
+import { RootStackNavigationProp } from '@/navigations';
 
 type HomeIconProps = {
   id: number;
