+++ src/components/molecules/Modal/ReviewDetailsModal.tsx
@@ -2,11 +2,11 @@ import useModalHook from '@hooks/modalHook';
 import BasicTwoButtonModal from '@components/atoms/Modal/BasicTwoButtonsModal';
 import BasicModalText from '@components/atoms/Text/BasicModalText';
 import { useNavigation } from '@react-navigation/native';
-import { RootStackNavigationProp } from '@/types/NavigationTypes';
+import { MyPageStackNavigationProp } from '@/navigations/MainBottomNavTab/MyPage/MyPageStack';
 
 function ReviewDetailsModal() {
   const { resetModal } = useModalHook();
-  const navigation = useNavigation<RootStackNavigationProp>();
+  const navigation = useNavigation<MyPageStackNavigationProp>();
 
   const onConfirm = () => {
     resetModal();
